requirements1.txt
Flask==2.2.2
numpy
Pillow
requests
tqdm
Werkzeug==2.2.2
